signature-valn-api/function_app.py
import azure.functions as func
import logging
import json
import requests
from typing import Dict, Any
import os
from datetime import datetime

app = func.FunctionApp()

@app.function_name(name="SignatureValidation")
@app.route(route="validate")
def signature_validation(req: func.HttpRequest) -> func.HttpResponse:
    """
    Azure Function to validate signatures and update Identifi documents
    """
    logging.info('Python HTTP trigger function processed a request.')
    
    try:
        # Get request body
        req_body = req.get_json()
        
        # Extract document information
        document_id = req_body.get('document_id')
        document_url = req_body.get('document_url')
        document_type = req_body.get('document_type')
        
        if not all([document_id, document_url, document_type]):
            return func.HttpResponse(
                json.dumps({"error": "Missing required fields: document_id, document_url, document_type"}),
                status_code=400,
                mimetype="application/json"
            )
        
        # Process signature validation
        validation_result = process_signature_validation(document_id, document_url, document_type)
        
        # Update Identifi document attributes
        update_result = update_identifi_document(document_id, validation_result)
        
        return func.HttpResponse(
            json.dumps({
                "status": "success",
                "document_id": document_id,
                "validation_result": validation_result,
                "update_result": update_result,
                "timestamp": datetime.utcnow().isoformat()
            }),
            status_code=200,
            mimetype="application/json"
        )
        
    except Exception as e:
        logging.error(f"Error processing request: {str(e)}")
        return func.HttpResponse(
            json.dumps({"error": f"Internal server error: {str(e)}"}),
            status_code=500,
            mimetype="application/json"
        )

def process_signature_validation(document_id: str, document_url: str, document_type: str) -> Dict[str, Any]:
    """
    Process signature validation for a document
    """
    # Mock validation logic - replace with actual signature validation
    validation_result = {
        "signatures_found": 2,
        "signatures_validated": 2,
        "confidence_score": 0.85,
        "validation_status": "valid",
        "processing_time": "1.2s"
    }
    
    logging.info(f"Signature validation completed for document {document_id}")
    return validation_result

def update_identifi_document(document_id: str, validation_result: Dict[str, Any]) -> Dict[str, Any]:
    """
    Update Identifi document with validation results
    """
    try:
        # Identifi API endpoint (replace with actual endpoint)
        identifi_api_url = os.getenv("IDENTIFI_API_URL", "https://api.identifi.com")
        api_key = os.getenv("IDENTIFI_API_KEY")
        
        if not api_key:
            logging.warning("IDENTIFI_API_KEY not configured")
            return {"status": "skipped", "reason": "API key not configured"}
        
        # Prepare update payload
        update_payload = {
            "document_id": document_id,
            "attributes": {
                "signature_validation_status": validation_result["validation_status"],
                "signature_confidence_score": validation_result["confidence_score"],
                "signatures_found": validation_result["signatures_found"],
                "signatures_validated": validation_result["signatures_validated"],
                "last_validation_date": datetime.utcnow().isoformat()
            },
            "notes": f"Signature validation completed: {validation_result['validation_status']} with {validation_result['confidence_score']} confidence"
        }
        
        # Make API call to Identifi
        headers = {
            "Authorization": f"Bearer {api_key}",
            "Content-Type": "application/json"
        }
        
        response = requests.post(
            f"{identifi_api_url}/documents/{document_id}/update",
            json=update_payload,
            headers=headers,
            timeout=30
        )
        
        if response.status_code == 200:
            logging.info(f"Successfully updated Identifi document {document_id}")
            return {"status": "success", "response": response.json()}
        else:
            logging.error(f"Failed to update Identifi document {document_id}: {response.status_code}")
            return {"status": "failed", "response_code": response.status_code, "response": response.text}
            
    except Exception as e:
        logging.error(f"Error updating Identifi document {document_id}: {str(e)}")
        return {"status": "error", "error": str(e)}

@app.function_name(name="BulkSignatureValidation")
@app.route(route="bulk-validate")
def bulk_signature_validation(req: func.HttpRequest) -> func.HttpResponse:
    """
    Azure Function for bulk signature validation
    """
    logging.info('Bulk signature validation function processed a request.')
    
    try:
        req_body = req.get_json()
        document_batch = req_body.get('documents', [])
        
        if not document_batch:
            return func.HttpResponse(
                json.dumps({"error": "No documents provided in batch"}),
                status_code=400,
                mimetype="application/json"
            )
        
        results = []
        for document in document_batch:
            try:
                # Process individual document
                validation_result = process_signature_validation(
                    document.get('document_id'),
                    document.get('document_url'),
                    document.get('document_type')
                )
                
                # Update Identifi
                update_result = update_identifi_document(
                    document.get('document_id'),
                    validation_result
                )
                
                results.append({
                    "document_id": document.get('document_id'),
                    "status": "success",
                    "validation_result": validation_result,
                    "update_result": update_result
                })
                
            except Exception as e:
                results.append({
                    "document_id": document.get('document_id'),
                    "status": "error",
                    "error": str(e)
                })
        
        return func.HttpResponse(
            json.dumps({
                "status": "completed",
                "total_documents": len(document_batch),
                "results": results,
                "timestamp": datetime.utcnow().isoformat()
            }),
            status_code=200,
            mimetype="application/json"
        )
        
    except Exception as e:
        logging.error(f"Error processing bulk validation: {str(e)}")
        return func.HttpResponse(
            json.dumps({"error": f"Bulk validation failed: {str(e)}"}),
            status_code=500,
            mimetype="application/json"
        )
